Don't count abandoned jobs as job queue pressure
It looks like the reason these jobs arn't processing is because
abandoned jobs are being considered as part of the acquired jobs
count, so if a job gets abandoned it keeps taking up a slot in
our job pressure calculation. These shouldn't count as pressure
because they are not running anymore.

Change-Id: I44fbce2b7dc47345ab0e3745d1653f418d75943d