Support class=pageimage for designating a page image
This is similar to the existing class=notpageimage functionality which explicitly excludes an image from becoming the page image. However, in this case, we add 1000 to the existing score rather than simply returning 1000, so that in the case of multiple images with this class on a page, ties can be broken easily.

Bug: T91683
Change-Id: I38241511f20b0acd71efccfb5b32bebe69bd94e9